US322 [Service] Calculate Other Services - Singapore Insurance changed commission type from Integer to Float

+++ apac-services-exchange-order-module/apac-services-exchange-order-api/src/main/java/com/cwt/bpg/cbt/exchange/order/model/Insurance.java
@@ -17,13 +17,13 @@ public class Insurance {
 	private String type;
 	
 	@NotNull
-	private Integer commission;
+	private Float commission;
 
-	public Integer getCommission() {
+	public Float getCommission() {
 		return commission;
 	}
 
-	public void setCommission(Integer commission) {
+	public void setCommission(Float commission) {
 		this.commission = commission;
 	}
